stomp 초기세팅, 패키지 구조 설정
