+++ src/main/java/com/springboot/jwttest/chat/model/dto/ChatDto.java
@@ -1,0 +1,10 @@
+package com.springboot.jwttest.chat.model.dto;
+
+import lombok.Getter;
+
+@Getter
+public class ChatDto {
+    private Long chatRoomId;
+    private String sender;
+    private String message;
+}
