+++ src/main/java/com/springboot/jwttest/chat/model/vo/ChatRoom.java
@@ -1,0 +1,4 @@
+package com.springboot.jwttest.chat.model.vo;
+
+public class ChatRoom {
+}
