+++ src/main/java/com/springboot/jwttest/jwt/auth/controller/UserController.java
@@ -1,9 +1,7 @@
 package com.springboot.jwttest.jwt.auth.controller;
 
-import com.springboot.jwttest.jwt.auth.dto.MeRes;
-import com.springboot.jwttest.user.mapper.UserMapper;
-import com.springboot.jwttest.user.repository.UserRepository;
-import com.springboot.jwttest.user.vo.User;
+import com.springboot.jwttest.jwt.auth.dto.MeResonseDto;
+import com.springboot.jwttest.user.model.repository.UserRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
@@ -17,17 +15,16 @@ import org.springframework.web.bind.annotation.RestController;
 @RequiredArgsConstructor
 public class UserController {
 
-    private final UserMapper userMapper;
     private final UserRepository userRepository;
 
     @GetMapping("/me")
-    public ResponseEntity<MeRes> me(Authentication authentication) {
+    public ResponseEntity<MeResonseDto> me(Authentication authentication) {
         if(authentication == null || authentication.getPrincipal() == null) {
             return ResponseEntity.status(HttpStatus.UNAUTHORIZED).build();
         }
         Integer userId = (Integer) authentication.getPrincipal();
         return userRepository.findById(userId)
-                .map(user -> ResponseEntity.ok(MeRes.from(user)))
+                .map(user -> ResponseEntity.ok(MeResonseDto.from(user)))
                 .orElseGet(() -> ResponseEntity.status(HttpStatus.NOT_FOUND).build());
     }
 }
