--- src/main/java/com/springboot/jwttest/jwt/auth/dto/LoginReq.java
@@ -1,8 +1,0 @@
-package com.springboot.jwttest.jwt.auth.dto;
-
-import jakarta.validation.constraints.NotBlank;
-
-public record LoginReq(
-        @NotBlank String login,   // 로그인 ID 또는 이메일
-        @NotBlank String password
-) {}
