+++ src/main/java/com/springboot/jwttest/jwt/auth/dto/LoginRequestDto.java
@@ -1,0 +1,8 @@
+package com.springboot.jwttest.jwt.auth.dto;
+
+import jakarta.validation.constraints.NotBlank;
+
+public record LoginRequestDto(
+        @NotBlank String login,   // 로그인 ID 또는 이메일
+        @NotBlank String password
+) {}
