--- src/main/java/com/springboot/jwttest/jwt/auth/dto/TokenRes.java
@@ -1,7 +1,0 @@
-package com.springboot.jwttest.jwt.auth.dto;
-
-public record TokenRes(
-        String accessToken,
-        String refreshToken,
-        long   expiresInSec
-) {}
