+++ src/main/java/com/springboot/jwttest/jwt/auth/dto/TokenResponseDto.java
@@ -1,0 +1,7 @@
+package com.springboot.jwttest.jwt.auth.dto;
+
+public record TokenResponseDto(
+        String accessToken,
+        String refreshToken,
+        long   expiresInSec
+) {}
