--- src/main/java/com/springboot/jwttest/jwt/token/mapper/RefreshTokenMapper.java
@@ -1,18 +1,0 @@
-package com.springboot.jwttest.jwt.token.mapper;
-
-import org.apache.ibatis.annotations.Mapper;
-import org.apache.ibatis.annotations.Param;
-
-import java.time.LocalDateTime;
-
-@Mapper
-public interface RefreshTokenMapper {
-
-//    int upsert(@Param("userId") int userId,
-//               @Param("tokenHash") String tokenHash,
-//               @Param("expiresAt") LocalDateTime expiresAt);
-//
-//    String selectHashByUserId(@Param("userId") int userId);
-//
-//    int deleteByUserId(@Param("userId") int userId);
-}
