--- src/main/java/com/springboot/jwttest/post/dto/PostRequestDto.java
@@ -1,9 +1,0 @@
-package com.springboot.jwttest.post.dto;
-
-import lombok.Getter;
-
-@Getter
-public class PostRequestDto {
-    private String title;
-    private String content;
-}
