--- src/main/java/com/springboot/jwttest/post/mappers/postMapper.java
@@ -1,9 +1,0 @@
-package com.springboot.jwttest.post.mappers;
-
-import com.springboot.jwttest.post.vo.Post;
-import org.apache.ibatis.annotations.Mapper;
-
-@Mapper
-public interface postMapper {
-    int insert(Post post);
-}
