+++ src/main/java/com/springboot/jwttest/post/model/dto/PostRequestDto.java
@@ -1,0 +1,9 @@
+package com.springboot.jwttest.post.model.dto;
+
+import lombok.Getter;
+
+@Getter
+public class PostRequestDto {
+    private String title;
+    private String content;
+}
