+++ src/main/java/com/springboot/jwttest/post/model/mappers/postMapper.java
@@ -1,0 +1,9 @@
+package com.springboot.jwttest.post.model.mappers;
+
+import com.springboot.jwttest.post.model.vo.Post;
+import org.apache.ibatis.annotations.Mapper;
+
+@Mapper
+public interface postMapper {
+    int insert(Post post);
+}
