+++ src/main/java/com/springboot/jwttest/post/model/repository/PostRepository.java
@@ -1,0 +1,7 @@
+package com.springboot.jwttest.post.model.repository;
+
+import com.springboot.jwttest.post.model.vo.Post;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface PostRepository extends JpaRepository<Post,Integer> {
+}
