+++ src/main/java/com/springboot/jwttest/user/model/repository/UserRepository.java
@@ -1,0 +1,10 @@
+package com.springboot.jwttest.user.model.repository;
+
+import com.springboot.jwttest.user.model.vo.User;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.Optional;
+
+public interface UserRepository extends JpaRepository<User, Integer> {
+    Optional<User> findById(String login);
+}
