+++ src/main/java/com/springboot/jwttest/user/model/vo/Sport.java
@@ -1,0 +1,4 @@
+package com.springboot.jwttest.user.model.vo;
+
+public class Sport {
+}
