--- src/main/resources/mapper/PostMapper.xml
@@ -1,18 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.springboot.jwttest.post.mappers.postMapper">
-    <resultMap id="PostMap" type="Post">
-        <id property="postId" column="post_id"/>
-        <result property="userId" column="user_id" />
-        <result property="title" column="title" />
-        <result property="content" column="content" />
-        <result property="createdAt" column="created_at" />
-        <result property="updatedAt" column="updated_at" />
-    </resultMap>
-    <insert id="insert" parameterType="Post" useGeneratedKeys="true" keyProperty="postId">
-        INSERT INTO posts(user_id, title, content, created_at, updated_at)
-        VALUES (#{userId}, #{title}, #{content}, NOW(), NOW())
-    </insert>
-</mapper>
