--- src/main/resources/mapper/RefreshTokenMapper.xml
@@ -1,22 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.springboot.jwttest.jwt.token.mapper.RefreshTokenMapper">
-
-    <insert id="upsert">
-        INSERT INTO refresh_token(user_id, token_hash, expires_at, created_at)
-        VALUES(#{userId}, #{tokenHash}, #{expiresAt}, NOW())
-        ON DUPLICATE KEY UPDATE
-        token_hash = VALUES(token_hash),
-        expires_at = VALUES(expires_at)
-    </insert>
-
-    <select id="selectHashByUserId" parameterType="int" resultType="string">
-        SELECT token_hash FROM refresh_token WHERE user_id = #{userId}
-    </select>
-
-    <delete id="deleteByUserId" parameterType="int">
-        DELETE FROM refresh_token WHERE user_id = #{userId}
-    </delete>
-</mapper>
