--- src/main/resources/mapper/UserMapper.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.springboot.jwttest.user.mapper.UserMapper">
-
-    <resultMap id="userMap" type="User">
-        <id     property="userId"     column="user_id"/>
-        <result property="profileImage" column="profile_image"/>
-        <result property="phoneNumber"  column="phone_number"/>
-        <result property="dmOption"     column="dm_option"/>
-        <result property="createdAt"    column="created_at"/>
-        <result property="updatedAt"    column="updated_at"/>
-    </resultMap>
-
-    <select id="selectByLogin" parameterType="string" resultMap="userMap">
-        SELECT * FROM user
-        WHERE id = #{login} OR email = #{login}
-        LIMIT 1
-    </select>
-
-    <select id="selectByUserId" parameterType="int" resultMap="userMap">
-        SELECT * FROM user WHERE user_id = #{userId}
-    </select>
-
-    <insert id="insertUser" parameterType="User" useGeneratedKeys="true" keyProperty="userId">
-        INSERT INTO user
-        (id, email, password, profile_image, name, nickname, gender, age,
-        address, phone_number, dm_option, status, created_at, updated_at)
-        VALUES
-        (#{id}, #{email}, #{password}, #{profileImage}, #{name}, #{nickname}, #{gender}, #{age},
-        #{address}, #{phoneNumber}, #{dmOption}, #{status}, NOW(), NOW())
-    </insert>
-</mapper>
